wf-details: lazy load ROOT file preview
Lazy load ROOT file preview so that `jsroot` is not part of the main
application bundle.
